requirements.txt
openai-whisper==20231117
torch==2.2.2
torchvision==0.17.2
transformers==4.40.1
tqdm==4.66.2
pytesseract==0.3.13
Pillow==11.1.0
opencv-python==4.11.0.86
pdfplumber==0.11.5
python-pptx==1.0.2
openpyxl==3.1.5
docx2txt==0.9
pandas==2.2.3
sentence-transformers==2.7.0
ftfy==6.2.0
regex==2024.5.15
scipy==1.13.1
faiss-cpu==1.8.0   
